refactor: 카드사 라벨 상수를 전역 상수로 수정
